% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_NSW.R
\docType{data}
\name{divisions_NSW}
\alias{divisions_NSW}
\title{Allhomes NSW suburb (division) names and IDs}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5271 rows and 5 columns.
}
\usage{
data(divisions_NSW)
}
\description{
Allhomes NSW suburb (division) names and IDs
}
\keyword{datasets}
