% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ACT_UV_data.R
\name{get_ACT_UV_data}
\alias{get_ACT_UV_data}
\title{Get unimproved value (UV) data from the ACT Revenue Office.}
\usage{
get_ACT_UV_data(suburb, quiet = FALSE)
}
\arguments{
\item{suburb}{A \code{character} scalar denoting an ACT suburb.}

\item{quiet}{If \code{TRUE} then messages are suppressed.}
}
\value{
A \code{tibble}
}
\description{
Get unimproved value (UV) data from the ACT Revenue Office.
}
