% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NSW_UV_data.R
\name{get_NSW_UV_data}
\alias{get_NSW_UV_data}
\title{Get unimproved value (UV) data from the NSW Valuer General.}
\usage{
get_NSW_UV_data(suburb, keep = TRUE, quiet = TRUE)
}
\arguments{
\item{suburb}{A \code{character} scalar denoting a NSW suburb.}

\item{keep}{A \code{logical} scalar denoting whether to keep UV data file. If
\code{TRUE} store in working directory.}

\item{quiet}{If \code{TRUE} then messages are suppressed.}
}
\value{
A \code{tibble}
}
\description{
Get unimproved value (UV) data from the NSW Valuer General.
}
