% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_past_sales_data.R
\name{get_past_sales_data}
\alias{get_past_sales_data}
\title{Extract historical Allhomes past sales data for an ACT/NSW suburb.}
\usage{
get_past_sales_data(suburb, year, max_entries = 5000)
}
\arguments{
\item{suburb}{A \code{character} scalar denoting a specific suburb. Format must
be "<suburb_name>, <state/territory_abbreviation>", e.g. "Balmain, ACT".}

\item{year}{An \code{integer} scalar denoting the year. If \code{NULL}, then \emph{all}
historical data for \code{suburb} are returned. Otherwise sales data for the
indicated \code{year} are returned.}

\item{max_entries}{The maximum number of records returned. Defaults to 5000.}
}
\value{
A \code{tibble}.
}
\description{
Extract historical Allhomes past sales data for an ACT/NSW suburb.
}
\examples{
\donttest{
get_past_sales_data(
    "Balmain, NSW",
    2020L)
}
}
