% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_localities_act.R
\docType{data}
\name{localities_act}
\alias{localities_act}
\title{Allhomes ACT suburb and localities data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 129 rows and 4 columns.
}
\source{
\href{allhomes.com.au}{allhomes.com.au}
}
\usage{
data(localities_act)
}
\description{
Allhomes ACT suburb and localities data
}
\keyword{datasets}
