% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_past_sales_data.R
\name{tokenise_column}
\alias{tokenise_column}
\title{Tokenise column data}
\usage{
tokenise_column(df, col, col_names)
}
\arguments{
\item{df}{A \code{tibble}.}

\item{col}{A column as unquoted expression.}

\item{col_names}{An \code{character} vector with column names.}
}
\value{
A \code{tibble}.
}
\description{
Tokenise column data
}
\keyword{internal}
