% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_ACT.R
\docType{data}
\name{divisions_ACT}
\alias{divisions_ACT}
\title{Allhomes ACT suburb (division) names and IDs}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 126 rows and 6 columns.
}
\usage{
data(divisions_ACT)
}
\description{
Allhomes ACT suburb (division) names and IDs
}
\keyword{datasets}
