% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_past_sales_data.R
\name{extract_past_sales_data}
\alias{extract_past_sales_data}
\title{Extract Allhomes past sales data for a single suburb + ID + year.}
\usage{
extract_past_sales_data(division_name, division_id, year, quiet = FALSE)
}
\arguments{
\item{division_name}{A \code{character} string denoting the Allhomes division
name (suburb).}

\item{division_id}{An \code{integer} scalar denoting the Allhomes division ID.}

\item{year}{An \code{integer} scalar denoting the year.}

\item{quiet}{If \code{TRUE} then messages are suppressed. Currently ignored.}
}
\value{
A \code{tibble}.
}
\description{
Extract Allhomes past sales data for a suburb + ID + year. This is an
internal function that gets called by vectorised \code{get_past_sales_data()}.
This function is currently exported but this may change in the future.
}
\examples{
\dontrun{
extract_past_sales_data("ainslie", 14743, 2021)
}
}
