% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_past_sales_data.R
\name{get_past_sales_data}
\alias{get_past_sales_data}
\title{Get Allhomes past sales data}
\usage{
get_past_sales_data(suburb, id, year, quiet = FALSE)
}
\arguments{
\item{suburb}{A \code{character} string denoting the suburb.}

\item{id}{An \code{integer} scalar denoting the Allhomes locality ID}

\item{year}{An \code{integer} scalar denoting the year.}

\item{quiet}{If \code{TRUE} then messages are suppressed.}
}
\value{
A \code{tibble}.
}
\description{
Get Allhomes past sales data
}
\examples{
\dontrun{
get_past_sales_data("ainslie", 14743, 2021)
}
}
